refactor: Club 비즈니스 로직 패키지 변경 및 주석 처리
