--- src/main/java/com/mobble/mobbleserver/application/club/core/port/provided/ClubUpdatePort.java
@@ -1,9 +1,0 @@
-package com.mobble.mobbleserver.application.club.core.port.provided;
-
-import com.mobble.mobbleserver.infrastructure.web.club.core.dto.request.ClubRequestDto;
-import com.mobble.mobbleserver.infrastructure.web.club.core.dto.response.ClubResponseDto;
-
-public interface ClubUpdatePort {
-
-    ClubResponseDto updateClub(Long clubId, Long memberId, ClubRequestDto dto);
-}
