+++ src/main/java/com/mobble/mobbleserver/application/club/port/provided/ClubDeletePort.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.application.club.core.port.provided;
+package com.mobble.mobbleserver.application.club.port.provided;
 
 public interface ClubDeletePort {
