+++ src/main/java/com/mobble/mobbleserver/application/club/port/provided/ClubQueryPort.java
@@ -1,0 +1,14 @@
+package com.mobble.mobbleserver.application.club.port.provided;
+
+import com.mobble.mobbleserver.infrastructure.web.club.dto.request.ClubSearchRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.club.dto.response.ClubResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.club.dto.response.ClubSummaryDto;
+
+import java.util.List;
+
+public interface ClubQueryPort {
+
+    ClubResponseDto findClubById(Long clubId, Long memberId);
+
+    List<ClubSummaryDto> searchClubs(ClubSearchRequestDto dto, Long memberId);
+}
