+++ src/main/java/com/mobble/mobbleserver/application/club/port/provided/ClubUpdatePort.java
@@ -1,0 +1,9 @@
+package com.mobble.mobbleserver.application.club.port.provided;
+
+import com.mobble.mobbleserver.infrastructure.web.club.dto.request.ClubRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.club.dto.response.ClubResponseDto;
+
+public interface ClubUpdatePort {
+
+    ClubResponseDto updateClub(Long clubId, Long memberId, ClubRequestDto dto);
+}
